Figured out the Serial issue
Fixed the Serial issue that wouldn't print. Can't use while (!Serial) in setup with Arduino Uno because the hardware is setup to emulate USB but doesn't actually have the same process. Different than a Teensy, so Serial is automatically connected and not an object that can be called on in the instance of (!Serial).