Added little face
Added a little face, getting the drawCircle to work. For some reason, the drawCircle function doesn't work well in the EVERY_N_MILLISECONDS function, and has to be updated outside on it's own.
The screen is very glitchy, oftentimes the white eyes shudder and the initial loading screen gets stuck in the background.